requirements.txt
# Dependencias generales
boto3==1.34.43
uuid==1.30
pyjwt==2.8.0

# Para diagramas AWS
diagrams==0.23.4

# Para diagramas ER
eralchemy2==1.3.6
sqlalchemy==2.0.25

# Para diagramas JSON
networkx==3.3
matplotlib==3.8.2
pydot==1.4.2

# Para procesamiento de imágenes
pillow==10.2.0
